refactor: split CLI command surfaces
