+++ plugins/codex-autoresearch/scripts/autoresearch.ts
@@ -8,9 +8,11 @@ import { createInterface } from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 import { fileURLToPath } from "node:url";
 import { buildDashboardViewModel, buildWatchdogSummary } from "../lib/dashboard-view-model.js";
-import { writeContextCapsule } from "../lib/context-capsule.js";
 import { createDashboardCommands } from "../lib/commands/dashboard.js";
 import { createInspectCommands } from "../lib/commands/inspect.js";
+import { createLaneRunnerCommand } from "../lib/commands/lane-runner.js";
+import { createPartialResultsCommand } from "../lib/commands/partial-results.js";
+import { createSessionForensicsCommand } from "../lib/commands/session-forensics.js";
 import { createCliCommandHandlers, runCliCommand } from "../lib/cli-handlers.js";
 import { buildDriftReport } from "../lib/drift-doctor.js";
 import { analyzeExperimentEconomics } from "../lib/experiment-economics.js";
@@ -27,15 +29,11 @@ import {
   defaultEvidenceStatusForRun,
 } from "../lib/evidence-registry.js";
 import { buildExperimentMemory } from "../lib/experiment-memory.js";
-import { mergeEvidenceClaims } from "../lib/evidence-index.js";
 import {
   finalizeCurrentTree as buildFinalizeCurrentTree,
   finalizePreview as buildFinalizePreview,
 } from "../lib/finalize-preview.js";
-import {
-  buildPartialResultEvidenceClaim,
-  discoverPartialResultCandidates,
-} from "../lib/partial-results.js";
+import { discoverPartialResultCandidates } from "../lib/partial-results.js";
 import { integrationsCommand } from "../lib/integrations.js";
 import {
   gapCandidates as buildGapCandidates,
@@ -52,7 +50,6 @@ import {
   revalidateRecipeCatalogProvenance,
 } from "../lib/recipes.js";
 import { serveAutoresearch } from "../lib/live-server.js";
-import { parseSessionForensics } from "../lib/session-forensics.js";
 import {
   parseMetricLines,
   runProcess as runBoundedProcess,
@@ -186,6 +183,54 @@ const { benchmarkLint, benchmarkInspect, checksInspect } = createInspectCommands
   resolveWorkDir,
   runShell,
   validateMetricName,
+});
+
+const sessionForensics = createSessionForensicsCommand({
+  boolOption,
+  pluginRoot: PLUGIN_ROOT,
+  positiveIntegerOption,
+  resolveWorkDir,
+  shellQuote,
+});
+
+const partialResultsCommand = createPartialResultsCommand({
+  appendJsonl,
+  assertFreshLastRunPacket,
+  boolOption,
+  computeConfidence,
+  currentState,
+  deleteLastRunPacket,
+  finiteMetric,
+  loopContinuation,
+  readConfig,
+  readLastRunPacket,
+  researchSlugFromArgs,
+  resolveWorkDir,
+});
+
+const laneRunner = createLaneRunnerCommand({
+  appendJsonl,
+  assertNoDirtyPathsOutsideWriteScope,
+  assertWriteScopeIntegrity,
+  boolOption,
+  buildParallelOrchestrationContext,
+  commandLooksMutating,
+  commandLooksUnsafeForWriteScope,
+  currentState,
+  dashboardSettings,
+  gitStatusPorcelain,
+  insideGitRepo,
+  latestLaneResults,
+  normalizeLaneMode,
+  normalizeParallelLane,
+  normalizeRelativePaths,
+  positiveIntegerOption,
+  resolveLaneWorktree,
+  resolveWorkDir,
+  runShell,
+  synthesizeLaneDecision,
+  tailText,
+  writeScopeSnapshot,
 });
 
 function usage() {
@@ -1824,74 +1869,6 @@ function canonicalTitle(kind: unknown, fallback: unknown): string {
     "next-packet": "Run the next measured packet",
   };
   return titles[text] || String(fallback || "Next action");
-}
-
-async function sessionForensics(args: LooseObject): Promise<LooseObject> {
-  const { workDir } = resolveWorkDir(args.working_dir || args.cwd);
-  const apply = boolOption(args.apply, false);
-  const dryRun = boolOption(args.dryRun, !apply);
-  const sessionJsonl = String(args.sessionJsonl || "");
-  const researchSlug = String(args.researchSlug || "");
-  const parsed = await parseSessionForensics({
-    sessionJsonl,
-    allowSnippets: boolOption(args.allowSnippets, false),
-    maxSnippets: positiveIntegerOption(args.maxSnippets, 8, "--max-snippets") ?? 8,
-    maxSnippetChars: positiveIntegerOption(args.maxSnippetChars, 320, "--max-snippet-chars") ?? 320,
-  });
-  if (!parsed.ok) {
-    return {
-      ...parsed,
-      wrote: false,
-    };
-  }
-  const capsule = await writeContextCapsule({
-    cwd: workDir,
-    researchSlug,
-    summary: parsed,
-    apply: apply && !dryRun,
-  });
-  const contextSignal = parsed.productSignals.find(
-    (signal: any) => signal.kind === "context_distillation_required",
-  );
-  const canonicalNextAction = contextSignal
-    ? {
-        kind: "context-distillation",
-        priority: 6,
-        reason: contextSignal.message,
-        command: `node ${shellQuote(path.join(PLUGIN_ROOT, "scripts", "autoresearch.mjs"))} session-forensics --cwd ${shellQuote(workDir)} --session-jsonl ${shellQuote(sessionJsonl)} --research-slug ${shellQuote(researchSlug)} --apply`,
-        triggeredBy: ["sessionForensics"],
-      }
-    : {
-        kind: "next-packet",
-        priority: 10,
-        reason: "Review imported signals, then continue with the safest next Autoresearch action.",
-        command: "",
-        triggeredBy: ["sessionForensics"],
-      };
-  return {
-    ok: true,
-    workDir,
-    dryRun: capsule.dryRun,
-    wrote: !capsule.dryRun,
-    outputDir: capsule.outputDir,
-    plannedFiles: capsule.files,
-    sourcePath: parsed.sourcePath,
-    timeWindow: parsed.timeWindow,
-    counts: parsed.counts,
-    responseCounts: parsed.responseCounts,
-    toolCounts: parsed.toolCounts,
-    commandClasses: parsed.commandClasses,
-    compactions: parsed.compactions,
-    goal: parsed.goal,
-    userCorrections: parsed.userCorrections,
-    productSignals: parsed.productSignals,
-    workflowWaste: parsed.workflowWaste,
-    blockers: parsed.blockers,
-    snippets: parsed.snippets,
-    evidenceClaims: capsule.evidenceIndex?.claims.length ?? null,
-    canonicalNextAction,
-    nextAction: canonicalNextAction.reason,
-  };
 }
 
 async function codexGoalBrief(args: LooseObject): Promise<LooseObject> {
@@ -5235,178 +5212,6 @@ function partialResultEligiblePacket(packet: LooseObject | null): boolean {
   return exitCode != null && exitCode !== 0;
 }
 
-async function partialResultsCommand(args: LooseObject) {
-  const { workDir } = resolveWorkDir(args.working_dir || args.cwd);
-  const state = currentState(workDir);
-  const artifact = args.artifact ? String(args.artifact) : "";
-  const recordId = args.record ? String(args.record).trim() : "";
-  const fromLast = boolOption(args.from_last ?? args.fromLast, !artifact || Boolean(recordId));
-  const lastRun = fromLast || recordId ? await readLastRunPacket(workDir) : null;
-  if (lastRun) await assertFreshLastRunPacket(workDir, lastRun);
-  const lastRunPacket =
-    lastRun ||
-    partialResultPacketFromArtifact({
-      artifact,
-      commandHash: args.command_hash ?? args.commandHash,
-      state,
-      workDir,
-    });
-  const discovery = await discoverPartialResultCandidates({
-    workDir,
-    primaryMetricName: state.config?.metricName || "metric",
-    lastRunPacket,
-  });
-  if (!recordId) {
-    return {
-      ok: true,
-      workDir,
-      source: lastRun ? "last-run" : "artifact",
-      candidates: discovery.candidates,
-      skippedArtifacts: discovery.skippedArtifacts,
-      nextAction: discovery.candidates.length
-        ? "Review a candidate, then record it as diagnostic measure evidence with --record <candidate-id>."
-        : "No partial-result candidates were found.",
-    };
-  }
-  if (!lastRun) {
-    throw new Error(
-      "--record requires a fresh last-run packet so salvaged evidence links to its source packet.",
-    );
-  }
-  const candidate = discovery.candidates.find((item: any) => item.id === recordId);
-  if (!candidate) throw new Error(`partial result candidate not found: ${recordId}`);
-  return await recordPartialResultCandidate({ workDir, state, lastRun, candidate, args });
-}
-
-function partialResultPacketFromArtifact({
-  artifact,
-  commandHash,
-  state,
-  workDir,
-}: LooseObject): LooseObject {
-  if (!artifact) throw new Error("--artifact is required unless --from-last is used.");
-  const artifactPath = path.isAbsolute(artifact) ? path.resolve(artifact) : artifact;
-  return {
-    ok: false,
-    workDir,
-    history: {
-      segment: state.segment,
-      config: lastRunConfigSnapshot(state.config),
-      nextRun: state.results.length + 1,
-    },
-    packetEvidence: {
-      packetId: `artifact-${createHash("sha256").update(String(artifactPath)).digest("hex").slice(0, 12)}`,
-      metricName: state.config?.metricName || "metric",
-      commandIdentity: {
-        commandHash: commandHash ? String(commandHash) : "",
-      },
-      artifacts: [
-        {
-          name: path.basename(String(artifactPath)) || "partial-result",
-          path: String(artifactPath),
-          exists: true,
-          quarantined: false,
-        },
-      ],
-    },
-  };
-}
-
-async function recordPartialResultCandidate({
-  workDir,
-  state,
-  lastRun,
-  candidate,
-  args,
-}: LooseObject) {
-  const metricName = candidate.metricName || state.config?.metricName || "metric";
-  const metric = finiteMetric(candidate.metricValue);
-  if (metric == null) {
-    throw new Error(
-      `partial result candidate ${candidate.id} has no finite metric and must stay manual-review only.`,
-    );
-  }
-  const sourcePacketId = lastRun?.packetEvidence?.packetId || "";
-  const researchSlug = researchSlugFromArgs({
-    slug: args.research_slug ?? args.researchSlug ?? "partial-results",
-  });
-  const evidenceClaim = buildPartialResultEvidenceClaim(candidate);
-  const evidenceIndex = await mergeEvidenceClaims(workDir, researchSlug, [evidenceClaim]);
-  const experiment: LooseObject = {
-    run: state.results.length + 1,
-    commit: "",
-    metric,
-    metrics: {
-      [metricName]: metric,
-    },
-    metricEligible: false,
-    status: "measure",
-    description:
-      args.description ||
-      `Diagnostic partial result from ${candidate.artifactName} row ${candidate.rowIndex}.`,
-    timestamp: Date.now(),
-    segment: state.segment,
-    confidence: null,
-    promotion: {
-      label: "measurement",
-      reasons: ["Recorded from a partial-result salvage candidate; diagnostic measure only."],
-    },
-    asi: {
-      hypothesis: "Recover diagnostic evidence from a partial benchmark artifact.",
-      evidence: `${metricName}=${metric} from ${candidate.artifactPath} row ${candidate.rowIndex}.`,
-      rollback_reason:
-        "Source packet crashed or timed out, so this evidence cannot be treated as promotion-grade.",
-      next_action_hint:
-        "Use this diagnostic row to choose the next packet; rerun fresh before promotion.",
-      partial_result: {
-        candidateId: candidate.id,
-        status: candidate.status,
-        reason: candidate.reason,
-        sourcePacketId,
-        artifactName: candidate.artifactName,
-        artifactPath: candidate.artifactPath,
-        rowIndex: candidate.rowIndex,
-        provenance: candidate.provenance,
-        evidenceClaimId: evidenceClaim.id,
-        researchSlug,
-      },
-      promotionGrade: false,
-    },
-    artifacts: {
-      [candidate.artifactName]: candidate.artifactPath,
-    },
-    partialResult: {
-      candidateId: candidate.id,
-      sourcePacketId,
-      evidenceClaimId: evidenceClaim.id,
-      researchSlug,
-      validationStatus: candidate.status,
-    },
-  };
-  experiment.confidence = computeConfidence(
-    [...state.current, experiment],
-    state.config?.bestDirection || "lower",
-  );
-  appendJsonl(workDir, experiment);
-  await deleteLastRunPacket(workDir);
-  const stateAfter = currentState(workDir);
-  return {
-    ok: true,
-    workDir,
-    experiment,
-    evidenceClaim,
-    evidenceIndex: {
-      slug: researchSlug,
-      claims: evidenceIndex.claims.length,
-    },
-    lastRunCleared: true,
-    baseline: stateAfter.baseline,
-    best: stateAfter.best,
-    confidence: stateAfter.confidence,
-    continuation: loopContinuation(workDir, stateAfter, readConfig(workDir), "logged"),
-  };
-}
-
 async function publicState(args: LooseObject): Promise<LooseObject> {
   const { workDir, config } = resolveWorkDir(args.working_dir || args.cwd);
   const state = currentState(workDir);
@@ -6289,171 +6094,6 @@ async function researchFanout(args: LooseObject) {
     dryRun,
     fanoutPlan: plan,
     parallelLanes: lanes,
-  };
-}
-
-async function laneRunner(args: LooseObject) {
-  const { workDir, config } = resolveWorkDir(args.working_dir || args.cwd);
-  const state = currentState(workDir);
-  const settings = dashboardSettings(config);
-  const { parallelLanes: lanes } = buildParallelOrchestrationContext({
-    workDir,
-    state,
-    config,
-    settings,
-  });
-  const laneId = String(
-    args.lane_id || args.laneId || args.lane || lanes[0]?.id || "read-only-scout",
-  );
-  const lane =
-    lanes.find((candidate: LooseObject) => candidate.id === laneId || candidate.label === laneId) ||
-    normalizeParallelLane({ id: laneId, label: laneId }, lanes.length, config);
-  const mode = normalizeLaneMode(args.mode, lane.mode);
-  const dryRun = boolOption(args.dry_run ?? args.dryRun, !boolOption(args.yes, false));
-  const command = String(args.command || "").trim();
-  const timeBudgetSeconds =
-    positiveIntegerOption(
-      args.time_budget_seconds ??
-        args.timeBudgetSeconds ??
-        args.timeout_seconds ??
-        args.timeoutSeconds,
-      300,
-      "--time-budget-seconds",
-    ) || 300;
-  const writeScope = normalizeRelativePaths(
-    args.write_scope ?? args.writeScope ?? args.commit_paths ?? args.commitPaths,
-    "--write-scope",
-  );
-  const worktreePath = String(
-    args.worktree_path || args.worktreePath || args.worktree || "",
-  ).trim();
-  if (mode === "implementation" && !worktreePath && writeScope.length === 0) {
-    throw new Error(
-      "Implementation lanes require explicit isolation: pass --worktree <path> or --write-scope <paths> before running.",
-    );
-  }
-  if (mode === "read_only_scout" && command && commandLooksMutating(command)) {
-    throw new Error("Read-only scout lanes cannot run commands that look mutating.");
-  }
-  const allowNonGitReadOnlyCommand = boolOption(
-    args.allow_non_git_command ?? args.allowNonGitCommand,
-    false,
-  );
-  if (mode === "read_only_scout" && command && !dryRun && !(await insideGitRepo(workDir))) {
-    if (!allowNonGitReadOnlyCommand) {
-      throw new Error(
-        "Read-only scout lanes cannot run commands outside a Git worktree without porcelain verification. Use --worktree or implementation mode with --write-scope for isolated edits, or pass --allow-non-git-command only when the command is provably read-only.",
-      );
-    }
-  }
-  if (
-    mode === "implementation" &&
-    !worktreePath &&
-    writeScope.length > 0 &&
-    command &&
-    commandLooksUnsafeForWriteScope(command)
-  ) {
-    throw new Error(
-      "Implementation lane --write-scope cannot run git cleanup, history, or stash commands in the main checkout; use a separate --worktree.",
-    );
-  }
-
-  const runCwd =
-    mode === "implementation" && worktreePath
-      ? await resolveLaneWorktree(workDir, worktreePath)
-      : workDir;
-  const beforeStatus =
-    mode === "read_only_scout" && command && !dryRun ? await gitStatusPorcelain(workDir) : null;
-  let writeScopeBefore: LooseObject | null = null;
-  if (mode === "implementation" && !worktreePath && writeScope.length > 0 && command && !dryRun) {
-    await assertNoDirtyPathsOutsideWriteScope(workDir, writeScope);
-    writeScopeBefore = await writeScopeSnapshot(workDir);
-  }
-  let commandResult: LooseObject | null = null;
-  if (command && !dryRun) {
-    const result = await runShell(command, runCwd, timeBudgetSeconds, {
-      retainMetricNames: [state.config.metricName || config.metricName || ""],
-    });
-    commandResult = {
-      code: result.exitCode,
-      timedOut: result.timedOut,
-      durationSeconds: result.durationSeconds,
-      output: tailText(result.output || "", 20, 4000),
-    };
-    if (beforeStatus != null) {
-      const afterStatus = await gitStatusPorcelain(workDir);
-      if (afterStatus !== beforeStatus) {
-        throw new Error(
-          "Read-only scout lane changed the git working tree; discard or isolate the change before continuing.",
-        );
-      }
-    }
-    if (writeScopeBefore) {
-      await assertWriteScopeIntegrity(workDir, writeScope, writeScopeBefore);
-    }
-  }
-
-  const explicitSummary = String(args.summary || "").trim();
-  const explicitRecommendation = String(
-    args.recommendation || args.next_action || args.nextAction || "",
-  ).trim();
-  const resultStatus =
-    args.result_status ||
-    args.resultStatus ||
-    (commandResult
-      ? commandResult.code === 0 && !commandResult.timedOut
-        ? "completed"
-        : "failed"
-      : explicitSummary || explicitRecommendation
-        ? "completed"
-        : "planned");
-  const commandSucceeded =
-    commandResult && Number(commandResult.code) === 0 && commandResult.timedOut !== true;
-  const evidenceAccepted = Boolean(
-    String(resultStatus).toLowerCase() === "completed" &&
-    (commandSucceeded || explicitSummary || explicitRecommendation),
-  );
-  const result = {
-    status: resultStatus,
-    summary:
-      explicitSummary || (commandResult ? "Lane command completed." : "Lane result recorded."),
-    recommendation: explicitRecommendation || lane.nextActionHint,
-    evidenceAccepted,
-    command: command || "",
-    timeBudgetSeconds,
-    isolation: {
-      mode,
-      worktree: worktreePath,
-      writeScope,
-    },
-    commandResult,
-  };
-  const entry = {
-    type: "lane_result",
-    timestamp: Date.now(),
-    segment: state.segment,
-    lane: {
-      id: lane.id,
-      title: lane.title || lane.label,
-      mode,
-    },
-    result,
-  };
-  const existingResults = latestLaneResults(workDir, state.segment);
-  const laneResults = dryRun ? existingResults : [...existingResults, entry];
-  const coordinatorRecommendation = synthesizeLaneDecision({
-    workDir,
-    laneResults,
-    fallbackLane: lane,
-  });
-  if (!dryRun) appendJsonl(workDir, entry);
-  return {
-    ok: true,
-    workDir,
-    dryRun,
-    lane: entry.lane,
-    result,
-    coordinatorRecommendation,
   };
 }
